Refactor of DetailScreen and NewsItem
- NewsItem is now an independent  file like rest of detail screens
- DetailScreen is now similar to IndexScreen only a bridge that renders the chosen direction
