+++ src/components/screens/Tour.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import { SafeAreaView, ScrollView, View } from 'react-native';
+import { View } from 'react-native';
 import _filter from 'lodash/filter';
 
 import { device, texts } from '../../config';
@@ -47,56 +47,54 @@ export const Tour = ({ data }) => {
     });
 
   return (
-    <SafeAreaView>
-      <ScrollView>
-        {!!images && images.length > 1 && <ImagesCarousel data={images} />}
-        {!!images && images.length === 1 && <Image source={images[0].picture} />}
+    <View>
+      {!!images && images.length > 1 && <ImagesCarousel data={images} />}
+      {!!images && images.length === 1 && <Image source={images[0].picture} />}
 
-        {!!title && (
-          <View>
-            <TitleContainer>
-              <Title>{title}</Title>
-            </TitleContainer>
-            {device.platform === 'ios' && <TitleShadow />}
-          </View>
-        )}
+      {!!title && (
+        <View>
+          <TitleContainer>
+            <Title>{title}</Title>
+          </TitleContainer>
+          {device.platform === 'ios' && <TitleShadow />}
+        </View>
+      )}
 
-        <Wrapper>
-          {!!logo && <Logo source={{ uri: logo }} />}
+      <Wrapper>
+        {!!logo && <Logo source={{ uri: logo }} />}
 
-          <InfoCard category={category} addresses={addresses} contact={contact} webUrls={webUrls} />
-        </Wrapper>
+        <InfoCard category={category} addresses={addresses} contact={contact} webUrls={webUrls} />
+      </Wrapper>
 
-        <TourCard lengthKm={lengthKm} addresses={addresses} />
+      <TourCard lengthKm={lengthKm} addresses={addresses} />
 
-        {!!description && (
-          <View>
-            <TitleContainer>
-              <Title>{texts.eventRecord.description}</Title>
-            </TitleContainer>
-            {device.platform === 'ios' && <TitleShadow />}
-            <Wrapper>
-              <HtmlView html={description} />
-            </Wrapper>
-          </View>
-        )}
+      {!!description && (
+        <View>
+          <TitleContainer>
+            <Title>{texts.eventRecord.description}</Title>
+          </TitleContainer>
+          {device.platform === 'ios' && <TitleShadow />}
+          <Wrapper>
+            <HtmlView html={description} />
+          </Wrapper>
+        </View>
+      )}
 
-        {!!operatingCompany && (
-          <View>
-            <TitleContainer>
-              <Title>{texts.eventRecord.operatingCompany}</Title>
-            </TitleContainer>
-            {device.platform === 'ios' && <TitleShadow />}
-            <OperatingCompanyInfo
-              name={operatingCompany.name}
-              address={operatingCompany.address}
-              contact={operatingCompany.contact}
-              webUrls={operatingCompany.contact.webUrls}
-            />
-          </View>
-        )}
-      </ScrollView>
-    </SafeAreaView>
+      {!!operatingCompany && (
+        <View>
+          <TitleContainer>
+            <Title>{texts.eventRecord.operatingCompany}</Title>
+          </TitleContainer>
+          {device.platform === 'ios' && <TitleShadow />}
+          <OperatingCompanyInfo
+            name={operatingCompany.name}
+            address={operatingCompany.address}
+            contact={operatingCompany.contact}
+            webUrls={operatingCompany.contact.webUrls}
+          />
+        </View>
+      )}
+    </View>
   );
 };
 /* eslint-enable complexity */
