+++ src/components/screens/index.js
@@ -1,6 +1,7 @@
 export * from './home/Service';
 export * from './EventRecord';
 export * from './InfoCard';
+export * from './NewsItem';
 export * from './PointOfInterest';
 export * from './PriceCard';
 export * from './OpeningTimesCard';
